delete bootstrap5 views
