--- src/Events/FieldBag/Flashed.php
@@ -1,34 +1,0 @@
-<?php
-
-namespace Digitlimit\Alert\Events\FieldBag;
-
-use Digitlimit\Alert\Types\FieldBag;
-use Illuminate\Broadcasting\Channel;
-use Illuminate\Broadcasting\InteractsWithSockets;
-use Illuminate\Broadcasting\PrivateChannel;
-use Illuminate\Foundation\Events\Dispatchable;
-use Illuminate\Queue\SerializesModels;
-
-class Flashed
-{
-    use Dispatchable, InteractsWithSockets, SerializesModels;
-
-    /**
-     * Create a new event instance.
-     */
-    public function __construct(
-        public FieldBag $alert
-    ) {}
-
-    /**
-     * Get the channels the event should broadcast on.
-     *
-     * @return array<int, Channel>
-     */
-    public function broadcastOn(): array
-    {
-        return [
-            new PrivateChannel('channel-name'),
-        ];
-    }
-}
