+++ src/Foundation/AbstractTheme.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Digitlimit\Alert\Themes;
+
+use Digitlimit\Alert\Contracts\ThemeInterface;
+
+abstract class AbstractTheme implements ThemeInterface {}
