--- src/Types/FieldBag.php
@@ -1,115 +1,0 @@
-<?php
-
-namespace Digitlimit\Alert\Types;
-
-use Digitlimit\Alert\Contracts\Taggable;
-use Digitlimit\Alert\Events\FieldBag\Flashed;
-use Digitlimit\Alert\Message\AbstractMessage;
-use Digitlimit\Alert\Message\MessageInterface;
-use Illuminate\Support\MessageBag;
-use Illuminate\Validation\Validator;
-use Digitlimit\Alert\Traits;
-
-class FieldBag extends AbstractMessage implements MessageInterface, Taggable
-{
-    use Traits\Taggable;
-
-    /**
-     * Field message back.
-     */
-    public MessageBag $messages;
-
-    /**
-     * Create a new field alert instance.
-     *
-     * @return void
-     */
-    public function __construct(
-        Validator|MessageBag|null $bag = null
-    ) {
-        parent::__construct();
-        $this->setBag($bag);
-    }
-
-    public function setBag(Validator|MessageBag|null $bag = null): self
-    {
-        if (is_a($bag, Validator::class)) {
-            $this->errors($bag);
-        } elseif (is_a($bag, MessageBag::class)) {
-            $this->messages($bag);
-        } else {
-            $this->messages = new MessageBag;
-        }
-
-        return $this;
-    }
-
-    /**
-     * Message store key for the field alert.
-     */
-    public function key(): string
-    {
-        return 'bag';
-    }
-
-    /**
-     * Set messages.
-     */
-    public function messages(MessageBag $messages): self
-    {
-        $this->messages = $messages;
-
-        return $this;
-    }
-
-    /**
-     * Set errors.
-     */
-    public function errors(Validator $validator): self
-    {
-        $this->messages = $validator->errors();
-
-        return $this;
-    }
-
-    /**
-     * Fetch a message for a given field name.
-     */
-    public function messageFor(string $name): string
-    {
-        return $this->messages->first($name);
-    }
-
-    /**
-     * Convert the field bag alert to an array.
-     */
-    public function toArray(): array
-    {
-        return array_merge(parent::toArray(), [
-            'type' => $this->key(),
-            'messages' => $this->messages,
-        ]);
-    }
-
-    /**
-     * Fill the field bag alert from an array.
-     */
-    public static function fill(array $alert): MessageInterface
-    {
-        $bag = new static;
-        $bag->id($alert['id']);
-        $bag->messages($alert['messages']);
-
-        return $bag;
-    }
-
-    /**
-     * Flash field instance to store.
-     */
-    public function flash(): void
-    {
-        parent::flash();
-
-        Flashed::dispatch($this);
-    }
-}
